Turn review.py script into a package
In preparation for some potential feature additions, turn this single script
into a package.
