+++ Pipfile
@@ -23,5 +23,5 @@ qa = "sh -c 'pipenv run checkformat && pipenv run typecheck && pipenv run lint'"
 checkformat = "black --check ."
 lint = "ruff check . --ignore E501"
 format = "black ."
-typecheck = "mypy --strict ."
-review = "python review.py"
+typecheck = "mypy --strict -m dependabot_batch_review"
+review = "python -m dependabot_batch_review"
